bug修复3.2： 单元测试1.1： 完成积分增加、减少，添加、查看标记，获取图片URL，查看已发布，已接受任务

+++ src/main/java/wnderful/imgannotator/util/FileHelper.java
@@ -97,10 +97,12 @@ public class FileHelper {
             if (dir.isDirectory()) {
                 String[] children = dir.list();
 
-                for (String aChildren : children) {
-                    boolean success = deleteAll(path + "/" + aChildren);
-                    if (!success) {
-                        return false;
+                if(children!=null){
+                    for (String aChildren : children) {
+                        boolean success = deleteAll(path + "/" + aChildren);
+                        if (!success) {
+                            return false;
+                        }
                     }
                 }
             }
